+++ app/profile/[id]/page.tsx
@@ -4,6 +4,7 @@ import { useQuery } from '@tanstack/react-query';
 
 import { LoadingArea, SettingIcon } from '@/components/atoms';
 import { GlobalNavigationBar, HeaderBar } from '@/components/molecules';
+import { ProfileProps } from '@/features/profile';
 import { MyProfile } from '@/features/profile/components/organisms/my-page';
 import { OtherPage } from '@/features/profile/components/organisms/other-page';
 import { css } from '@/styled-system/css';
@@ -12,19 +13,6 @@ import { flex } from '@/styled-system/patterns';
 export type Mypage = {
   params: { id: number };
 };
-
-export interface ProfileProps {
-  status: number;
-  code: string;
-  data: {
-    memberId: number;
-    nickname: string;
-    isMyProfile: boolean;
-    followerCount: number;
-    followingCount: number;
-    introduction: string;
-  };
-}
 
 const fetchProfileData = async (id: number) => {
   const response = await fetch(`/api/profile/${id}`);
