+++ features/profile/components/organisms/index.ts
@@ -1,3 +1,4 @@
 export * from './my-page';
 export * from './other-page';
+export * from './profile-container';
 export * from './profile-edit-form';
