+++ features/profile/types/profile.ts
@@ -1,0 +1,12 @@
+export interface ProfileProps {
+  status: number;
+  code: string;
+  data: {
+    memberId: number;
+    nickname: string;
+    isMyProfile: boolean;
+    followerCount: number;
+    followingCount: number;
+    introduction: string;
+  };
+}
